refactor(test-runs): extract each tab into a folder with its related components
