+++ galasa-ui/src/components/test-runs/table-design/TableDesignContent.tsx
