+++ galasa-ui/src/components/test-runs/table-design/TableDesignRow.tsx
